renamed video cards to project cards
